Moved `QueryComponentError` and `QueryEntityError` into their own Module
